fix combined Searchbar and SearchbarEntry
