+++ documentation/docs/searchbar.md
@@ -76,6 +76,14 @@ A react native style object which takes care of styling the container around the
 | :--------------------------: | :------------------------------------------------------------------------------------------------------------------------------------------------------------: | :-------------------: |
 | ViewStyle <img width="500"/> | `{flexDirection: 'row', padding: 12, backgroundColor: useColors().white, borderRadius: 39, borderWidth: 1, borderColor:useColors().font_4,}`<img width="500"/> | No <img width="500"/> |
 
+### editable
+
+boolean value to determine if the searchbar is editable or not. If `editable` is `false`, the searchbar acts like an entry to the search screen. The `onPress` prop must provide navigation to the search screen that will use `Searchbar` with `editable` set as `true`. When `editable` is `true`, the searchbar can be used like `TextInput` for the search logic.
+
+|            Type            |         Default         |       Required        |
+| :------------------------: | :---------------------: | :-------------------: |
+| boolean <img width="500"/> | None <img width="500"/> | No <img width="500"/> |
+
 ### inputStyle
 
 A react native style object which takes care of styling the text in the searchbar.
@@ -83,6 +91,23 @@ A react native style object which takes care of styling the text in the searchba
 |             Type             |                                                               Default                                                                |       Required        |
 | :--------------------------: | :----------------------------------------------------------------------------------------------------------------------------------: | :-------------------: |
 | TextStyle <img width="500"/> | `{flex: 4, paddingHorizontal: 5, paddingVertical: 0, fontSize: Fonts.size.regular_17, color: useColors().font_1}` <img width="500"/> | No <img width="500"/> |
+
+### leftIcon
+
+Name of the icon to be showed which appears in the left end of the searchbar. Only icons from **react-native-vector-icons/MaterialIcons** are supported.
+Note : If `editable` is `true` and `showBackIcon` is `true`, **only the `backIcon` will be rendered in the left end**.
+
+|           Type            |         Default         |       Required        |
+| :-----------------------: | :---------------------: | :-------------------: |
+| string <img width="500"/> | null <img width="500"/> | No <img width="500"/> |
+
+### leftLogo
+
+Logo to be displayed at the left end. **leftLogo has higher precedence than leftIcon** i.e., if both leftIcon and leftLogo are passed, only leftLogo is rendered in the left end (if `editable` is `false`).
+
+|                  Type                  |         Default         |       Required        |
+| :------------------------------------: | :---------------------: | :-------------------: |
+| ImageSourcePropType <img width="500"/> | null <img width="500"/> | No <img width="500"/> |
 
 ### onBackIconPress
 
@@ -94,7 +119,7 @@ Function to be called when the back icon is pressed. There will not be any funct
 
 ### onChangeText
 
-function to use the text entered into the searchbar. This is the search logic.
+Function to use the text entered into the searchbar. This is the search logic.
 
 |            Type             |         Default         |       Required        |
 | :-------------------------: | :---------------------: | :-------------------: |
@@ -107,6 +132,46 @@ Function to be called when the clear icon is pressed.
 |            Type             |                Default                |       Required        |
 | :-------------------------: | :-----------------------------------: | :-------------------: |
 | function <img width="500"/> | `onChangeText('')` <img width="500"/> | No <img width="500"/> |
+
+### onLeftIconPress
+
+Function to be called when the left icon is pressed.
+
+|            Type             |           Default            |       Required        |
+| :-------------------------: | :--------------------------: | :-------------------: |
+| function <img width="500"/> | `onPress` <img width="500"/> | No <img width="500"/> |
+
+### onLeftLogoPress
+
+Function to be called when the left logo is pressed.
+
+|            Type             |           Default            |       Required        |
+| :-------------------------: | :--------------------------: | :-------------------: |
+| function <img width="500"/> | `onPress` <img width="500"/> | No <img width="500"/> |
+
+### onRightIconPress
+
+Function to be called when the right icon is pressed.
+
+|            Type             |           Default            |       Required        |
+| :-------------------------: | :--------------------------: | :-------------------: |
+| function <img width="500"/> | `onPress` <img width="500"/> | No <img width="500"/> |
+
+### onRightLogoPress
+
+Function to be called when the right logo is pressed.
+
+|            Type             |           Default            |       Required        |
+| :-------------------------: | :--------------------------: | :-------------------: |
+| function <img width="500"/> | `onPress` <img width="500"/> | No <img width="500"/> |
+
+### onPress
+
+Function to be called when the searchbar with `editable` set as `false` is pressed. This should be mainly used to navigate to the search screen which will have the `Searchbar` component with `editable` set as `true`.
+
+|            Type             |         Default         |       Required        |
+| :-------------------------: | :---------------------: | :-------------------: |
+| function <img width="500"/> | null <img width="500"/> | No <img width="500"/> |
 
 ### placeholder
 
@@ -123,6 +188,23 @@ Color of placeholder
 |           Type            |                Default                |       Required        |
 | :-----------------------: | :-----------------------------------: | :-------------------: |
 | string <img width="500"/> | useColors().font_3 <img width="500"/> | No <img width="500"/> |
+
+### rightIcon
+
+Name of the icon to be showed which appears in the right end of the searchbar. Only icons from **react-native-vector-icons/MaterialIcons** are supported.
+Note : If `editable` is `true` and `showClearIcon` is `true`, **only the `clearIcon` will be rendered in the right end**.
+
+|           Type            |         Default         |       Required        |
+| :-----------------------: | :---------------------: | :-------------------: |
+| string <img width="500"/> | null <img width="500"/> | No <img width="500"/> |
+
+### rightLogo
+
+Logo to be displayed at the right end. **rightLogo has higher precedence than rightIcon** i.e., if both rightIcon and rightLogo are passed, only rightLogo is rendered in the right end (if `editable` is `false`).
+
+|                  Type                  |         Default         |       Required        |
+| :------------------------------------: | :---------------------: | :-------------------: |
+| ImageSourcePropType <img width="500"/> | null <img width="500"/> | No <img width="500"/> |
 
 ### showBackIcon
 
@@ -142,7 +224,7 @@ If `true`, `clearIcon` will be rendered.
 
 ### value
 
-Value to show in the searchbar. This field is needed to use the value entered by the user and apply search logic.
+Value to show in the searchbar. When `editable` is `true`, this prop is used for the search logic and for the text to be displayed in the searchbar.
 
 |           Type            |           Default            |       Required        |
 | :-----------------------: | :--------------------------: | :-------------------: |
