+++ src/components/Searchbar/helper.tsx
@@ -11,17 +11,11 @@ const getStyles = (Colors: ColorType) => {
       borderRadius: 39,
       borderWidth: 1,
       borderColor: Colors.font_4,
-    },
-    placeholderText: {
-      fontSize: Fonts.size.regular_17,
-      color: Colors.font_3,
-      paddingHorizontal: 5,
-      paddingVertical: 0,
-      flex: 4,
+      maxHeight: 54,
     },
     textInput: {
       flex: 4,
-      paddingHorizontal: 5,
+      paddingHorizontal: 10,
       paddingVertical: 0,
       fontSize: Fonts.size.regular_17,
       color: Colors.font_1,
