+++ src/components/index.tsx
@@ -12,7 +12,6 @@ import EmptyState from './EmptyState';
 import EmptyView from './EmptyView';
 import Switch from './Switch';
 import Searchbar from './Searchbar';
-import SearchbarEntry from './SearchbarEntry';
 import Tag from './Tag';
 
 export {
@@ -27,7 +26,6 @@ export {
   Ripple,
   Switch,
   Searchbar,
-  SearchbarEntry,
   Table,
   Tag,
   Input,
